requirements.txt
playwright
pytest
allure-pytest
PyYAML